fix: use method to get the request url
* fix: use method to get the request url

If you want to connect the plugins to a staging, it is possible to adjust the API_BASE_URL (in the plugin). Only an adjustment has to be made in the sdk to actually get the label from staging.

(cherry picked from commit faa22851d1aa6bb93619c6481092a4afea31515e)